app/main.py
from dotenv import load_dotenv
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

load_dotenv()

from app.api.routes import router as api_router

app = FastAPI(title="Phase 0 RAG API")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # dev ke liye
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(api_router)


@app.get("/")
def root() -> dict[str, str]:
    return {"message": "Phase 0 RAG API is running", "health": "/health"}


@app.get("/health")
def health_check() -> dict[str, str]:
    return {"status": "ok"}
